no close button
The "close" button is no longer needed. Previously, after the user clicked Confirm, there was no way to change what tiers were used, so we provided a Close button that would let them start over from a new page load. Now if the user decides to change what buttons are selected and clicks Confirm again, the result will change accordingly.

It would still be nice to be able to collapse the format window, but we shouldn't do that via a page refresh.